Implemented manual Log Enter + sharedPreferences
Added a new button, view, and functionality for manual log entering. This view is required to use deep links with Voice Assistant later, as it will be the view that will be opened.
Also, implemented sharedPreferences to store user data. Users don't need to relogin every time they enter the application, AND now if user logged in - the addLog activity will have all the required user data for proper functionaling (userId + userActivities)